Alterações 25/01 - Todo mundo odeia permissões

+++ src/Controller/CoresController.php
@@ -19,6 +19,11 @@ class CoresController extends AppController {
 	}
 
 	public function add() {
+		if($this->userObj->role < C_RoleTudo) {
+			$this->Flash->error(__('Você não possui permissão para realizar esta ação, contate um administrador.'));
+			return $this->redirect(['action' => 'index']);
+		}
+
 		$core = $this->Cores->newEmptyEntity();
 		if ($this->request->is('post')) {
 			$core = $this->Cores->patchEntity($core, $this->request->getData());
@@ -36,9 +41,13 @@ class CoresController extends AppController {
 	}
 
 	public function edit($id = null) {
-		$core = $this->Cores->get($id, [
-			'contain' => [],
-		]);
+		if($this->userObj->role < C_RoleTudo) {
+			$this->Flash->error(__('Você não possui permissão para realizar esta ação, contate um administrador.'));
+			return $this->redirect(['action' => 'index']);
+		}
+
+		$core = $this->Cores->get($id);
+		
 		if ($this->request->is(['patch', 'post', 'put'])) {
 			$core = $this->Cores->patchEntity($core, $this->request->getData());
 
@@ -55,6 +64,11 @@ class CoresController extends AppController {
 	}
 
 	public function delete($id = null) {
+		if($this->userObj->role < C_RoleTudo) {
+			$this->Flash->error(__('Você não possui permissão para realizar esta ação, contate um administrador.'));
+			return $this->redirect(['action' => 'index']);
+		}
+
 		$core = $this->Cores->get($id);
 
 		if ($this->Cores->delete($core)) {
