+++ src/Controller/ResponsibleController.php
@@ -12,14 +12,22 @@ class ResponsibleController extends AppController {
 	}
 
 	public function view($id = null) {
-		$responsibles = $this->Responsible->findyId($id)->first();
+		$responsible = $this->Responsible->findById($id)
+			->contain(['Users' => ['fields' => ['name', 'id']]])
+		->first();
 
 		$this->set(compact('responsible'));
 		$this->set('title', 'Visualizar responsável');
 	}
 
 	public function add() {
+		if($this->userObj->role < C_RoleTudo) {
+			$this->Flash->error(__('Você não possui permissão para realizar esta ação, contate um administrador.'));
+			return $this->redirect(['action' => 'index']);
+		}
+
 		$responsible = $this->Responsible->newEmptyEntity();
+		
 		if ($this->request->is('post')) {
 			$responsible = $this->Responsible->patchEntity($responsible, $this->request->getData());
 
@@ -30,15 +38,22 @@ class ResponsibleController extends AppController {
 
 			$this->Flash->error(__('Não foi possível salvar o responsável, tente novamente.'));
 		}
+		
+		$users = $this->Users->find('list', ['keyField' => 'id', 'valueField' => 'name'])->where(['type' => C_RoleResponsável])->order(['name ASC'])->toArray();
 
+		$this->set('users', $users);
 		$this->set(compact('responsible'));
 		$this->set('title', 'Cadastrar responsável');
 	}
 
 	public function edit($id = null) {
-		$responsible = $this->Responsible->get($id, [
-			'contain' => [],
-		]);
+		if($this->userObj->role < C_RoleTudo) {
+			$this->Flash->error(__('Você não possui permissão para realizar esta ação, contate um administrador.'));
+			return $this->redirect(['action' => 'index']);
+		}
+
+		$responsible = $this->Responsible->get($id);
+
 		if ($this->request->is(['patch', 'post', 'put'])) {
 			$responsible = $this->Responsible->patchEntity($responsible, $this->request->getData());
 			
@@ -49,12 +64,20 @@ class ResponsibleController extends AppController {
 
 			$this->Flash->error(__('Não foi possível salvar o responsável, tente novamente.'));
 		}
+		
+		$users = $this->Users->find('list', ['keyField' => 'id', 'valueField' => 'name'])->where(['type' => C_RoleResponsável])->order(['name ASC'])->toArray();
 
+		$this->set('users', $users);
 		$this->set(compact('responsible'));
 		$this->set('title', 'Alterar responsável');
 	}
 
 	public function delete($id = null) {
+		if($this->userObj->role < C_RoleTudo) {
+			$this->Flash->error(__('Você não possui permissão para realizar esta ação, contate um administrador.'));
+			return $this->redirect(['action' => 'index']);
+		}
+		
 		$responsible = $this->Responsible->get($id);
 
 		if ($this->Responsible->delete($responsible)) {
